UI Changes
Finished Text and Image buttons.
Removed radio buttons options.
Renamed Create A Message to Create A Segment.
Created new panels for Text and Image.
Color now works for all Segments.
All segments now clear the textbox when switched. (added to both the panel and the label)
All segments now fill the text box with any prevous entered text for that segment. (added to both the panel and the label)
Changed the mouse hover effect on the close button to DimGray. So it can be seen while the button is active.
All options now show up from the beginning. (not including Special Effects)
Remove the ability for the user to enter text into the ComboBoxes.