Added Support for 24 Segments
-This fixed the bug with all segments above 18.
-Fixed BackToMenu Button.
-Fixed form size when BackToMenu button is pressed.
-Play and pause button now disappear when BackToMenu button is pressed.
-Play button not appearing fixed.

Notes from last commit.
Segments and addSegmentButtons are now all made in a for loop and stored in a array. This means they no longer have a name so they are refereenced in the array.  There all use the same eventhandler which test which control the button is over. This allows me to control all the segments in for loops.